+++ Bugsnag/KSCrash/Source/KSCrash/Recording/BSG_KSCrashC.c
@@ -243,5 +243,8 @@ char *bsg_kscrash_captureThreadTrace(int discardDepth, int frameCount, uintptr_t
         bsg_kscrashsentry_resume_threads_user(false);
     }
     return trace;
-    
+}
+
+char *bsg_kscrash_captureAppForegroundStats() {
+    return bsg_kscrw_i_captureAppForegroundStats(crashContext());
 }