+++ Bugsnag/KSCrash/Source/KSCrash/Recording/BSG_KSCrashC.h
@@ -173,6 +173,11 @@ BSG_KSCrash_Context *crashContext(void);
  */
 char *bsg_kscrash_captureThreadTrace(int discardDepth, int frameCount, uintptr_t *callstack);
 
+/**
+ * Captures application foreground state info and returns it as a JSON string.
+ */
+char *bsg_kscrash_captureAppForegroundStats();
+
 #ifdef __cplusplus
 }
 #endif